+++ flutter_crud_app/lib/screens/users_detail_screen.dart
@@ -1,0 +1,79 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
+
+import '../data/user_data_service.dart';
+
+class UsuarioDetail extends HookWidget {
+  final String titulo;
+  final Usuario? usuarioAtual;
+
+  const UsuarioDetail({
+    Key? key, 
+    required this.titulo,
+    this.usuarioAtual}
+    ) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final nomeController =
+        useTextEditingController(text: usuarioAtual?.nome ?? '');
+    final emailController =
+        useTextEditingController(text: usuarioAtual?.email ?? '');
+    final cpfController =
+        useTextEditingController(text: usuarioAtual?.cpf ?? '');
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(titulo),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            TextField(
+              controller: nomeController,
+              decoration: InputDecoration(
+                labelText: AppLocalizations.of(context)!.userFieldName,
+              ),
+            ),
+            TextField(
+              controller: emailController,
+              decoration: InputDecoration(
+                labelText: AppLocalizations.of(context)!.userFieldEmail,
+              ),
+            ),
+            TextField(
+              controller: cpfController,
+              decoration: InputDecoration(
+                labelText: AppLocalizations.of(context)!.userFieldCpf,
+              ),
+            ),
+            const SizedBox(height: 16.0),
+            ElevatedButton(
+              onPressed: () {
+                if (nomeController.text.isNotEmpty &&
+                    emailController.text.isNotEmpty &&
+                    cpfController.text.isNotEmpty) {
+                  Usuario newUser = Usuario(
+                    nome: nomeController.text,
+                    email: emailController.text,
+                    cpf: cpfController.text,
+                  );
+                  Navigator.pop(context, newUser);
+                } else {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    SnackBar(
+                      content: Text(AppLocalizations.of(context)!.userAviso)
+                    ),
+                  );
+                }
+              },
+              child: Text(AppLocalizations.of(context)!.userSave),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
