+++ flutter_crud_app/lib/screens/users_list_screen.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 
-import 'widgets/user_card.dart';
-import 'widgets/my_app_bar.dart';
-import 'data/user_data_service.dart';
+import '../widgets/user_card.dart';
+import '../widgets/my_app_bar.dart';
+import '../data/user_data_service.dart';
+import 'users_detail_screen.dart';
+
 
 class UserScreen extends StatelessWidget {
   const UserScreen({super.key});
@@ -40,11 +41,12 @@ class UserScreen extends StatelessWidget {
                         return UsuarioCard(
                           cardTitle: value['dataObjects'][index].nome,
                           cardSubtitle: value['dataObjects'][index].email,
+                          
                           onEditPressed: () async {
                             Usuario? newUser = await Navigator.push(
                               context,
                               MaterialPageRoute(
-                                builder: (context) => UsuarioCadastro(
+                                builder: (context) => UsuarioDetail(
                                   titulo: AppLocalizations.of(context)!.userTitleEdit,
                                   usuarioAtual: value['dataObjects'][index]
                                 ),
@@ -58,9 +60,9 @@ class UserScreen extends StatelessWidget {
                               );
                             }
                           },
+                          
                           onDeletePressed: () {
-                            userDataService
-                                .deleteUser(value['dataObjects'][index]);
+                            userDataService.deleteUser(value['dataObjects'][index]);
                           },
                         );
                       }
@@ -86,7 +88,7 @@ class UserScreen extends StatelessWidget {
           Usuario? newUser = await Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => UsuarioCadastro(
+              builder: (context) => UsuarioDetail(
                 titulo: AppLocalizations.of(context)!.userTitleCreate),
             ),
           );
@@ -96,76 +98,6 @@ class UserScreen extends StatelessWidget {
           }
         },
         child: const Icon(Icons.add),
-      ),
-    );
-  }
-}
-
-class UsuarioCadastro extends HookWidget {
-  final String titulo;
-  final Usuario? usuarioAtual;
-
-  const UsuarioCadastro({Key? key, required this.titulo, this.usuarioAtual})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final nomeController =
-        useTextEditingController(text: usuarioAtual?.nome ?? '');
-    final emailController =
-        useTextEditingController(text: usuarioAtual?.email ?? '');
-    final cpfController =
-        useTextEditingController(text: usuarioAtual?.cpf ?? '');
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(titulo),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            TextField(
-              controller: nomeController,
-              decoration: InputDecoration(
-                labelText: AppLocalizations.of(context)!.userFieldName,
-              ),
-            ),
-            TextField(
-              controller: emailController,
-              decoration: InputDecoration(
-                labelText: AppLocalizations.of(context)!.userFieldEmail,
-              ),
-            ),
-            TextField(
-              controller: cpfController,
-              decoration: InputDecoration(
-                labelText: AppLocalizations.of(context)!.userFieldCpf,
-              ),
-            ),
-            const SizedBox(height: 16.0),
-            ElevatedButton(
-              onPressed: () {
-                if (nomeController.text.isNotEmpty &&
-                    emailController.text.isNotEmpty &&
-                    cpfController.text.isNotEmpty) {
-                  Usuario newUser = Usuario(
-                    nome: nomeController.text,
-                    email: emailController.text,
-                    cpf: cpfController.text,
-                  );
-                  Navigator.pop(context, newUser);
-                } else {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                        content: Text(AppLocalizations.of(context)!.userAviso)),
-                  );
-                }
-              },
-              child: Text(AppLocalizations.of(context)!.userSave),
-            ),
-          ],
-        ),
       ),
     );
   }
